Using built in nice number method
Removing method that was implemented in this code
