+++ include/general_functions.h
@@ -16,7 +16,5 @@ QColor GenerateRandomColor();
 QColor styleSheetToColor(QString stylesheet);
 int NETCDF_ERR(int status);
 bool isConnectedToNetwork();
-double niceNumber(double number, bool round);
-QVector<double> niceLabels(double inMin, double inMax, int nTicks, int &nFrac);
 
 #endif // GENERAL_FUNCTIONS_H
