+++ src/general_functions.cpp
@@ -194,52 +194,3 @@ bool isConnectedToNetwork(){
 
     return result;
 }
-
-double niceNumber(double number, bool round)
-{
-    int exp  = qFloor(qLn(number)/qLn(10.0));
-    double f = number / qPow(10,exp);
-    if(round)
-    {
-        if(f<1.5)
-            return qPow(10,exp);
-        else if(f<3.0)
-            return 2*qPow(10,exp);
-        else if(f<7.0)
-            return 5*qPow(10,exp);
-        else
-            return 10*qPow(10,exp);
-    }
-    else
-    {
-        if(f<=1.0)
-            return qPow(10,exp);
-        else if(f<=2.0)
-            return 2*qPow(10,exp);
-        else if(f<=5.0)
-            return 5*qPow(10,exp);
-        else
-            return 10*qPow(10,exp);
-    }
-}
-
-
-QVector<double> niceLabels(double inMin, double inMax, int nTicks, int &nFrac)
-{
-    double range    = niceNumber(qMax(qAbs(inMax),qAbs(inMin)),false);
-    double d        = niceNumber(range/(nTicks-1),false);
-    double graphMin = qFloor(inMin/d)*d;
-    double graphMax = qCeil(inMax/d)*d;
-    nFrac = qMax(-qFloor(qLn(d)/qLn(10)),0);
-
-    int idx = 0;
-    QVector<double> output;
-    for(double i=graphMin;i<graphMax+0.5*d;i=i+d)
-    {
-        idx = idx + 1;
-        output.resize(idx);
-        output[idx-1] = i;
-    }
-    return output;
-}
-
