+++ src/user_timeseries.cpp
@@ -46,84 +46,6 @@ user_timeseries::user_timeseries(QTableWidget *inTable, QCheckBox *inXAxisCheck,
     chart      = inChart;
 }
 
-
-int user_timeseries::plotData()
-{
-    int i,j,ierr,nFrac;
-    double TempValue,TempDate;
-    double unitConversion,addX,addY;
-    double ymin,ymax;
-    QVector<double> labels;
-    QVector<QLineSeries *> series;
-    QColor seriesColor;
-    QDateTime minDate,maxDate;
-
-    ierr = this->setMarkerID();
-    ierr = this->getDataBounds(ymin,ymax,minDate,maxDate);
-    labels = niceLabels(ymin,ymax,5,nFrac);
-    for(i=0;i<labels.length();i++)
-    {
-        if(labels[i]>ymax)
-            ymax = labels[i];
-        if(labels[i]<ymin)
-            ymin = labels[i];
-    }
-
-    QChart *thisChart = new QChart();
-    thisChart->setAnimationOptions(QChart::SeriesAnimations);
-    thisChart->legend()->setAlignment(Qt::AlignBottom);
-
-    QDateTimeAxis *axisX = new QDateTimeAxis;
-    axisX->setTickCount(5);
-    if(minDate.daysTo(maxDate)>90)
-        axisX->setFormat("MM/yyyy");
-    else if(minDate.daysTo(maxDate)>4)
-        axisX->setFormat("MM/dd/yyyy");
-    else
-        axisX->setFormat("MM/dd/yyyy hh:mm");
-    axisX->setTitleText("Date");
-    axisX->setMin(minDate);
-    axisX->setMax(maxDate);
-    thisChart->addAxis(axisX, Qt::AlignBottom);
-
-    QValueAxis *axisY = new QValueAxis;
-    axisY->setTickCount(5);
-    axisY->setTitleText(this->yLabelEdit->text());
-    axisY->setMin(ymin);
-    axisY->setMax(ymax);
-    thisChart->addAxis(axisY, Qt::AlignLeft);
-
-    series.resize(fileDataUnique.length());
-
-    for(i=0;i<fileDataUnique.length();i++)
-    {
-      series[i] = new QLineSeries();
-      series[i]->setName(table->item(i,1)->text());
-      seriesColor.setNamedColor(table->item(i,2)->text());
-      series[i]->setPen(QPen(seriesColor,3,Qt::SolidLine,Qt::RoundCap,Qt::RoundJoin));
-      unitConversion = table->item(i,3)->text().toDouble();
-      addX = table->item(i,4)->text().toDouble();
-      addY = table->item(i,5)->text().toDouble();
-
-      for(j=0;j<fileDataUnique[i].station[this->markerID].NumSnaps;j++)
-      {
-          TempDate = fileDataUnique[i].station[this->markerID].date[j].toMSecsSinceEpoch()+addX*3.6e+6;
-          TempValue = fileDataUnique[i].station[this->markerID].data[j]*unitConversion+addY;
-          if(TempValue>MOV_NULL_TS)
-            series[i]->append(TempDate,TempValue);
-      }
-      thisChart->addSeries(series[i]);
-      series[i]->attachAxis(axisX);
-      series[i]->attachAxis(axisY);
-    }
-
-    thisChart->setTitle(plotTitle->text());
-    chart->setRenderHint(QPainter::Antialiasing);
-    chart->setChart(thisChart);
-
-    return 0;
-}
-
 int user_timeseries::getDataBounds(double &ymin, double &ymax, QDateTime &minDate, QDateTime &maxDate)
 {
     int i,j;
